+++ numerical/linear-system/norm.cpp
@@ -6,7 +6,7 @@ using namespace std;
 double taxicab_norm(const vector<double> &u, const vector<double> &v)
 {
   double norm = 0;
-  for (size_t i = 0; i < u.size(); i++)
+  for (int i = 0; i < u.size(); i++)
   {
     norm += abs(u[i] - v[i]);
   }
@@ -16,7 +16,7 @@ double taxicab_norm(const vector<double> &u, const vector<double> &v)
 double max_norm(const vector<double> &u, const vector<double> &v)
 {
   double norm = 0, r;
-  for (size_t i = 0; i < u.size(); i++)
+  for (int i = 0; i < u.size(); i++)
   {
     r = abs(u[i] - v[i]);
     if (r > norm)
@@ -30,7 +30,7 @@ double max_norm(const vector<double> &u, const vector<double> &v)
 double euclid_norm(const vector<double> &u, const vector<double> &v)
 {
   double norm = 0;
-  for (size_t i = 0; i < u.size(); i++)
+  for (int i = 0; i < u.size(); i++)
   {
     norm += pow(u[i] - v[i], 2);
   }
